+++ app/Http/Controllers/PermissionController.php
@@ -12,12 +12,18 @@ use Yajra\DataTables\DataTables;
 class PermissionController extends Controller
 {
     public function index(){
+        if(!auth()->user()->can('view-permission')){
+            abort(403,'Unauthorized Action');
+        }
         $ninja = Auth::user();
         $permission = Permission::all();
         return view('permission.index',compact('permission','ninja'));
     }
 
     public function create(){
+        if(!auth()->user()->can('create-permission')){
+            abort(403,'Unauthorized Action');
+        }
         return view('permission.create');
     }
 
@@ -35,8 +41,17 @@ class PermissionController extends Controller
         return DataTables::of($permission)
 
             ->addColumn('action',function ($each){
-                $edit_icon = '<a href="'.route('permission.edit',$each->id).'" class="btn btn-warning btn-sm mr-2"><i class="fas fa-edit"></i></a>';
-                $delete_icon = '<a href="#" class="btn btn-danger btn-sm delete-btn"  data-id="'. $each->id .'"><i class="fas fa-trash-alt"></i></a>';
+
+                $edit_icon = '';
+                $delete_icon = '';
+                if(auth()->user()->can('edit-permission')){
+                    $edit_icon = '<a href="'.route('permission.edit',$each->id).'" class="btn btn-warning btn-sm mr-2"><i class="fas fa-edit"></i></a>';
+                }
+
+                if(auth()->user()->can('delete-permission')){
+                    $delete_icon = '<a href="#" class="btn btn-danger btn-sm delete-btn"  data-id="'. $each->id .'"><i class="fas fa-trash-alt"></i></a>';
+                }
+
                 return '<div>'.$edit_icon.$delete_icon.'</div>';
             })
             ->addColumn('plus-icon',function ($each){
@@ -47,6 +62,9 @@ class PermissionController extends Controller
     }
 
     public function edit($id){
+        if(!auth()->user()->can('edit-permission')){
+            abort(403,'Unauthorized Action');
+        }
         $permission = Permission::find($id);
         return view('permission.edit',compact('permission'));
     }
@@ -62,6 +80,9 @@ class PermissionController extends Controller
     }
 
     public function destroy($id){
+        if(!auth()->user()->can('delete-permission')){
+            abort(403,'Unauthorized Action');
+        }
         $permission = Permission::findOrFail($id);
         $permission->delete();
